backend/requirements.txt
# backend/requirements.txt

# Core Framework
FastAPI==0.104.1
uvicorn[standard]==0.24.0

# AI & ML Libraries
langchain==0.1.0
langchain-google-vertexai==1.0.0  # Google Vertex AI integration
sentence-transformers==2.2.2
chromadb==0.4.18

# Google Cloud Platform
google-cloud-aiplatform==1.38.0
google-cloud-firestore==2.13.1
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.4
google-cloud-monitoring==2.16.0
google-cloud-tasks==2.19.3
google-cloud-functions==1.20.4
google-cloud-secret-manager==2.24.0
google-cloud-error-reporting==3.6.0    # Phase E: Enhanced error tracking
google-cloud-memcache==0.1.0           # Phase E: Memorystore Memcached

# Gemini AI for v2.0 Strategic Scoring
google-generativeai==0.3.2             # Gemini LLM API
PyYAML==6.0.1                          # YAML configuration loading

# Data Processing
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
scipy==1.11.4
chromadb==0.4.18

# Export Libraries
reportlab==4.0.7
openpyxl==3.1.2

# Additional dependencies
pydantic==2.5.0                  # Data validation and serialization
httpx==0.25.2                    # Modern HTTP client
python-multipart==0.0.6          # File upload handling
aiohttp==3.9.1                   # Async HTTP client
beautifulsoup4==4.12.2           # HTML parsing
lxml==4.9.3                      # XML/HTML parser
requests==2.31.0                 # HTTP requests
psutil==7.1.0                    # System and process utilities

# Phase E: Advanced Orchestration & Observability
redis==5.0.1                     # Redis client for L2 cache
msgpack==1.0.7                   # Fast serialization for cache
python-memcached==1.62           # Memcached client (fallback)

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
pytest-html==4.1.1
pytest-json-report==1.5.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Pydantic settings
pydantic-settings==2.10.1
